// modules/it/maintenance-tools/maintenance-history/types/maintenance-history.ts
export interface MaintenanceHistory {

  id: number;

  tanggal: string;

  kode_pc: string;

  ruangan: string;

  teknisi: string;

  keluhan: string;

  keyboard_mouse: string;

  motherboard: string;

  ram: string;

  hardisk: string;

  power_supply: string;

  cleaner: string;

  baut: string;

  keterangan: string;

}